requirements.txt
altgraph==0.17.2
bcrypt==4.0.0
bdist-mpkg==0.5.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==36.0.2
dnslib==0.9.20
dnspython==2.2.1
Flask==2.2.2
future==0.18.2
idna==3.3
impacket==0.10.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
ldapdomaindump==0.9.3
macholib==1.16
MarkupSafe==2.1.1
modulegraph==0.19.2
paramiko==2.11.0
progress==1.6
py2app==0.28.2
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.15.0
pyftpdlib==1.5.6
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scapy==2.4.5
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1